Move common arangod code to separate repo
Bump go-helper version
