+++ cmd/cmd.go
@@ -49,7 +49,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/api"
 	deploymentApi "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 	"github.com/arangodb/kube-arangodb/pkg/crd"
-	"github.com/arangodb/kube-arangodb/pkg/deployment/agency/cache"
+	agencyConfig "github.com/arangodb/kube-arangodb/pkg/deployment/agency/config"
 	"github.com/arangodb/kube-arangodb/pkg/deployment/features"
 	"github.com/arangodb/kube-arangodb/pkg/deployment/reconcile"
 	"github.com/arangodb/kube-arangodb/pkg/generated/clientset/versioned/scheme"
@@ -237,7 +237,7 @@ func init() {
 	if err := features.Init(&cmdMain); err != nil {
 		panic(err.Error())
 	}
-	if err := cache.Init(&cmdMain); err != nil {
+	if err := agencyConfig.Init(&cmdMain); err != nil {
 		panic(err.Error())
 	}
 	if err := reconcile.ActionsConfigGlobal.Init(&cmdMain); err != nil {
