+++ pkg/deployment/agency/cache.go
@@ -27,10 +27,10 @@ import (
 
 	"github.com/rs/zerolog"
 
+	agencyCache "github.com/arangodb-helper/go-helper/pkg/arangod/agency/cache"
 	"github.com/arangodb-helper/go-helper/pkg/arangod/conn"
 
 	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
-	agencyCache "github.com/arangodb/kube-arangodb/pkg/deployment/agency/cache"
 	"github.com/arangodb/kube-arangodb/pkg/deployment/agency/state"
 	"github.com/arangodb/kube-arangodb/pkg/generated/metric_descriptions"
 	"github.com/arangodb/kube-arangodb/pkg/logging"
