--- pkg/deployment/agency/cache/interfaces.go
@@ -1,43 +1,0 @@
-//
-// DISCLAIMER
-//
-// Copyright 2023 ArangoDB GmbH, Cologne, Germany
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-
-package cache
-
-import (
-	"context"
-	"time"
-
-	"github.com/arangodb-helper/go-helper/pkg/arangod/conn"
-)
-
-type LeaderDiscovery interface {
-	Discover(ctx context.Context) (conn.Connection, error)
-}
-
-type StateLoader[T interface{}] interface {
-	State() (*T, uint64, bool)
-
-	Invalidate()
-	Valid() bool
-
-	UpdateTime() time.Time
-
-	Refresh(ctx context.Context, discovery LeaderDiscovery) error
-}
