+++ pkg/deployment/agency/config/config.go
@@ -18,7 +18,7 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 
-package cache
+package config
 
 import (
 	"time"
